refactor: refactors JWT handling for enhanced oop
Improves JWT management by renaming `JwtProvider` to `JwtManager` and introducing interfaces `JwtProvider` and `JwtValidator` for better separation of concerns.
This change enhances code maintainability and testability by decoupling token creation and validation logic.
Also adds logging for JWT validation failures to facilitate debugging.
