+++ src/main/java/com/lp/v2/security/jwt/JwtValidator.java
@@ -1,69 +1,7 @@
 package com.lp.v2.security.jwt;
 
-import io.jsonwebtoken.Claims;
-import io.jsonwebtoken.ExpiredJwtException;
-import io.jsonwebtoken.Jwts;
-import io.jsonwebtoken.MalformedJwtException;
-import io.jsonwebtoken.UnsupportedJwtException;
-import io.jsonwebtoken.security.SecurityException;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.stereotype.Component;
-
-import java.util.Date;
-
-@Component
-@RequiredArgsConstructor
-@Slf4j
-public class JwtValidator {
-
-    private final JwtKeyFactory jwtKeyFactory;
-
-    public String getSubject(String token) {
-        Claims claims = parseClaims(token);
-        return claims.getSubject();
-    }
-
-    public boolean validateToken(String token) {
-        try {
-            parseClaims(token);
-            log.debug("JWT token validation successful");
-            return true;
-        } catch (SecurityException | MalformedJwtException ex) {
-            log.warn("Invalid JWT signature: {}", ex.getMessage());
-            return false;
-        } catch (ExpiredJwtException ex) {
-            log.warn("Expired JWT token: {}", ex.getMessage());
-            return false;
-        } catch (UnsupportedJwtException ex) {
-            log.warn("Unsupported JWT token: {}", ex.getMessage());
-            return false;
-        } catch (IllegalArgumentException ex) {
-            log.warn("Invalid JWT token: {}", ex.getMessage());
-            return false;
-        } catch (Exception ex) {
-            log.error("Unexpected error during JWT validation: {}", ex.getMessage());
-            return false;
-        }
-    }
-
-    public Claims parseClaims(String token) {
-        return Jwts.parserBuilder()
-                .setSigningKey(jwtKeyFactory.getSecretKey())
-                .build()
-                .parseClaimsJws(token)
-                .getBody();
-    }
-
-    public boolean isTokenExpired(String token) {
-        try {
-            Claims claims = parseClaims(token);
-            return claims.getExpiration().before(new Date());
-        } catch (ExpiredJwtException ex) {
-            return true;
-        } catch (Exception ex) {
-            log.warn("Error checking token expiration: {}", ex.getMessage());
-            return true; // 예외 발생시 만료된 것으로 간주
-        }
-    }
+public interface JwtValidator {
+    boolean validateToken(String token);
+    boolean isTokenExpired(String token);
+    String getSubject(String token);
 }
